Split TweetSender into delivery mixins
sender.py had grown from 1349 to 2253 lines and held two methods over 300 lines
each. Move the forward, merged-forward, direct, helper and platform clusters
into mixins; sender.py keeps the constructor, the public entry points, the
low-level send calls and the ActionFailed classification (457 lines).

The ActionFailed checks stay put on purpose: the tests patch it as a module
global on delivery.sender, and moving the methods would resolve the name in the
new module and silently disable the patch.

Pure move: TweetSender keeps all 79 members with identical bodies.

Also drop the platform predicates and split helpers left unused after
PlatformResolver took over, along with the two platform constants they read.
