+++ delivery/sender.py
@@ -21,61 +21,48 @@ except ImportError:
     from astrbot.api.event import MessageChain
 
 try:
-    from astrbot.api.message_components import Plain
-except ImportError:
-    from astrbot.core.message.components import Plain
-
-try:
     from ..config import (
         configured_merge_tweet_threshold,
         resolve_send_image_attachments,
         resolve_send_video_attachments,
     )
-    from .lark_support import is_lark_platform
-    from ..shared import (
-        TweetItem,
-        safe_call,
-    )
-    from .default import DefaultDeliveryAdapter
-    from .onebot import OneBotDeliveryAdapter
-    from .outcomes import MergedSendOutcome, SendAttempt, SendOutcome
-    from .platforms import PlatformDeliveryRegistry, PlatformResolver, normalize_platform
-    from ..rendering import TweetBatch, TweetMessageRenderer
+    from ..shared import TweetItem
+    from .outcomes import SendAttempt, SendOutcome
+    from .platforms import PlatformDeliveryRegistry, PlatformResolver
+    from .sender_capabilities import SenderCapabilitiesMixin
+    from .sender_direct import SenderDirectMixin
+    from .sender_forward import SenderForwardMixin
+    from .sender_merged import SenderMergedForwardMixin
+    from .sender_helpers import SenderHelpersMixin
+    from ..rendering import TweetMessageRenderer
 except ImportError:
     from config import (
         configured_merge_tweet_threshold,
         resolve_send_image_attachments,
         resolve_send_video_attachments,
     )
-    from delivery.lark_support import is_lark_platform
-    from shared import (
-        TweetItem,
-        safe_call,
-    )
+    from shared import TweetItem
     from delivery import (
-        DefaultDeliveryAdapter,
-        MergedSendOutcome,
-        OneBotDeliveryAdapter,
         PlatformDeliveryRegistry,
         PlatformResolver,
         SendAttempt,
         SendOutcome,
-        normalize_platform,
     )
-    from rendering import TweetBatch, TweetMessageRenderer
+    from delivery.sender_capabilities import SenderCapabilitiesMixin
+    from delivery.sender_direct import SenderDirectMixin
+    from delivery.sender_forward import SenderForwardMixin
+    from delivery.sender_merged import SenderMergedForwardMixin
+    from delivery.sender_helpers import SenderHelpersMixin
+    from rendering import TweetMessageRenderer
 
 
-class TweetSender:
-    # AstrBot 的 Node/Nodes 合并转发主要由 OneBot v11 实现。
-    FORWARD_MESSAGE_PLATFORMS = {"aiocqhttp"}
-    QQ_DIRECT_VIDEO_SPLIT_PLATFORMS = FORWARD_MESSAGE_PLATFORMS | {
-        "qq",
-        "qq_official",
-        "qqofficial",
-        "onebot",
-        "onebot_v11",
-        "napcat",
-    }
+class TweetSender(
+    SenderCapabilitiesMixin,
+    SenderDirectMixin,
+    SenderForwardMixin,
+    SenderMergedForwardMixin,
+    SenderHelpersMixin,
+):
     FORWARD_TWEET_CHUNK_SIZE = 8
     # When NapCat/OneBot rejects a forward (often retcode 1200 / res_id fail),
     # recursively split the tweet list and retry smaller merges.
@@ -188,229 +175,6 @@ class TweetSender:
             and self._should_use_merge_for_count(tweet_count)
         )
 
-    async def _send_event_forward_chunks(
-        self,
-        event,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        notices: list[str] | None = None,
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-        on_delivered=None,
-    ) -> bool:
-        chunks = self._tweet_chunks(tweets)
-        indexed_chunks = []
-        index = tweet_start_index
-        for chunk in chunks:
-            indexed_chunks.append((index, chunk))
-            index += len(chunk)
-        return await self._send_chunked_bool(
-            indexed_chunks,
-            lambda item: self._send_event_forward_chunk(
-                event,
-                username,
-                instance,
-                item[1],
-                notices=notices,
-                tweet_start_index=item[0],
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-                on_delivered=on_delivered,
-            ),
-        )
-
-    async def _send_event_forward_chunk(
-        self,
-        event,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        notices: list[str] | None = None,
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-        on_delivered=None,
-    ) -> bool:
-        nodes = self.renderer.build_nodes(
-            event, username, instance, tweets, notices=notices,
-            start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-        raw_nodes = self.renderer.build_onebot_nodes(
-            event, username, instance, tweets, notices=notices,
-            start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-        try:
-            await event.send(event.chain_result([nodes]))
-            self._notify_delivered(on_delivered, len(tweets))
-            return True
-        except Exception as exc:
-            if self._is_uncertain_delivery_error(exc):
-                self._log_uncertain_delivery(
-                    "manual forwarded tweets", self._event_target(event), exc
-                )
-                self._notify_delivered(on_delivered, len(tweets))
-                return True
-            logger.warning(f"[NitterTweets] 发送合并转发节点失败: {exc}")
-
-        # 去掉视频后重试
-        if not media_only and any(
-            m.is_video for t in tweets for m in t.media if m.path
-        ):
-            try:
-                nodes_nv = self.renderer.build_nodes(
-                    event, username, instance, tweets,
-                    exclude_videos=True, notices=notices,
-                    start_index=tweet_start_index,
-                    media_only=media_only,
-                    omit_status_url=omit_status_url,
-                    hide_original_when_translated=hide_original_when_translated,
-                    link_style=link_style,
-                )
-                await event.send(event.chain_result([nodes_nv]))
-                logger.info("[NitterTweets] 初次失败后已发送去除视频的合并转发")
-                self._notify_delivered(on_delivered, len(tweets))
-                return True
-            except Exception as exc:
-                if self._is_uncertain_delivery_error(exc):
-                    self._log_uncertain_delivery(
-                        "manual tweets without videos", self._event_target(event), exc
-                    )
-                    self._notify_delivered(on_delivered, len(tweets))
-                    return True
-                logger.warning(
-                    f"[NitterTweets] 发送去除视频的合并转发节点失败: {exc}"
-                )
-
-        last_exc: Exception | None = None
-        try:
-            if await self._send_onebot_forward(event, raw_nodes):
-                self._notify_delivered(on_delivered, len(tweets))
-                return True
-            logger.warning(
-                f"[NitterTweets] 发送 OneBot 合并转发消息失败: action returned false "
-                f"(tweets={len(tweets)}, target={self._event_target(event)})"
-            )
-        except Exception as exc:
-            last_exc = exc
-            if self._is_uncertain_delivery_error(exc):
-                self._log_uncertain_delivery(
-                    "manual OneBot forward fallback", self._event_target(event), exc
-                )
-                self._notify_delivered(on_delivered, len(tweets))
-                return True
-            logger.warning(f"[NitterTweets] 发送 OneBot 合并转发消息失败: {exc}")
-
-        # retcode 1200 / res_id fail / explicit false: split smaller merges then retry.
-        # false return (no exception) is treated as payload reject so we can split.
-        should_split = last_exc is None or self._is_forward_payload_rejected_error(
-            last_exc
-        )
-        remaining = tweets
-        remaining_index = tweet_start_index
-        remaining_notices = notices
-        if should_split:
-            parts = self._split_tweets_for_forward_retry(tweets)
-            if parts:
-                logger.info(
-                    f"[NitterTweets] 合并转发失败，拆成 {len(parts)} 段重试 "
-                    f"(tweets={len(tweets)}, target={self._event_target(event)})"
-                )
-                index = tweet_start_index
-                for offset, part in enumerate(parts):
-                    part_delivered = 0
-
-                    def record_part_delivered(count: int) -> None:
-                        nonlocal part_delivered
-                        try:
-                            delivered = max(0, int(count))
-                        except (TypeError, ValueError, OverflowError):
-                            return
-                        delivered = min(len(part) - part_delivered, delivered)
-                        if delivered <= 0:
-                            return
-                        part_delivered += delivered
-                        self._notify_delivered(on_delivered, delivered)
-
-                    part_ok = await self._send_event_forward_chunk(
-                        event,
-                        username,
-                        instance,
-                        part,
-                        notices=notices if offset == 0 else None,
-                        tweet_start_index=index,
-                        media_only=media_only,
-                        omit_status_url=omit_status_url,
-                        hide_original_when_translated=hide_original_when_translated,
-                        link_style=link_style,
-                        on_delivered=record_part_delivered,
-                    )
-                    if part_ok and part_delivered < len(part):
-                        record_part_delivered(len(part) - part_delivered)
-                    if not part_ok:
-                        # Never re-send already-delivered parts as a full-batch fallback.
-                        remaining = list(part[part_delivered:])
-                        for later in parts[offset + 1 :]:
-                            remaining.extend(later)
-                        remaining_index = index + part_delivered
-                        remaining_notices = (
-                            notices if offset == 0 and part_delivered == 0 else None
-                        )
-                        break
-                    index += len(part)
-                else:
-                    return True
-
-        if not remaining:
-            return True
-
-        # Final fallback: direct send only the undelivered remainder.
-        logger.info(
-            f"[NitterTweets] 合并转发仍失败，降级直发 "
-            f"(tweets={len(remaining)}/{len(tweets)}, "
-            f"target={self._event_target(event)})"
-        )
-        try:
-            accepted = await self._delivery_adapter_for_event(event).send_event(
-                event,
-                username,
-                instance,
-                remaining,
-                notices=remaining_notices,
-                tweet_start_index=remaining_index,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            if accepted:
-                self._notify_delivered(on_delivered, len(remaining))
-            return accepted
-        except Exception as exc:
-            if self._is_uncertain_delivery_error(exc):
-                self._log_uncertain_delivery(
-                    "manual direct after forward fail", self._event_target(event), exc
-                )
-                self._notify_delivered(on_delivered, len(remaining))
-                return True
-            logger.warning(f"[NitterTweets] 合并转发降级直发仍失败: {exc}")
-            return False
-
     async def send_to_umo(
         self,
         context,
@@ -519,1203 +283,6 @@ class TweetSender:
     async def send_summary_to_umo(self, context, umo: str, summary: str) -> SendOutcome:
         return await self._delivery_adapter_for_umo(context, umo).send_summary_to_umo(
             context, umo, summary
-        )
-
-    async def _send_forward_chunks_to_umo(
-        self,
-        context,
-        umo: str,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        group_label: str = "",
-        header_text: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> SendOutcome:
-        chunks = self._tweet_chunks(tweets)
-        indexed_chunks = []
-        index = tweet_start_index
-        for chunk in chunks:
-            indexed_chunks.append((len(indexed_chunks), index, chunk))
-            index += len(chunk)
-        errors: list[str] = []
-        warnings: list[str] = []
-        delivery_errors: list[str] = []
-        delivered: list[str] = []
-        partial_delivery = False
-        for item in indexed_chunks:
-            outcome = await self._send_forward_chunk_to_umo(
-                context,
-                umo,
-                username,
-                instance,
-                item[2],
-                group_label=group_label if item[0] == 0 else "",
-                header_text=header_text if item[0] == 0 else "",
-                batch_summary=batch_summary if item[0] == 0 else "",
-                tweet_start_index=item[1],
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            if outcome.error:
-                errors.append(outcome.error)
-            if outcome.warning:
-                warnings.append(outcome.warning)
-            status = self._normalized_delivery_status(outcome)
-            if status in {"partial_failed", "failed"}:
-                partial_delivery = True
-            outcome_delivery_error = getattr(outcome, "delivery_error", "")
-            if outcome_delivery_error:
-                delivery_errors.append(outcome_delivery_error)
-            delivered.extend(
-                getattr(outcome, "delivered_status_ids", ())
-                or self._status_ids_from_tweets(item[2])
-                if outcome.success
-                else getattr(outcome, "delivered_status_ids", ())
-            )
-            if not outcome.success:
-                delivery_error = "; ".join(delivery_errors) or "; ".join(errors)
-                return SendOutcome(
-                    success=False,
-                    error="; ".join(errors) or outcome.error,
-                    warning="; ".join(warnings),
-                    delivery_status=(
-                        "partial_failed"
-                        if delivered or status == "partial_failed"
-                        else "failed"
-                    ),
-                    delivery_error=delivery_error or outcome.error,
-                    delivered_status_ids=self._dedupe_status_ids(delivered),
-                )
-        final_delivery_error = "; ".join(delivery_errors)
-        if partial_delivery and not final_delivery_error:
-            final_delivery_error = "; ".join(errors)
-        return SendOutcome(
-            success=True,
-            error="; ".join(errors),
-            warning="; ".join(warnings),
-            delivery_status="partial_failed" if partial_delivery else "success",
-            delivery_error=final_delivery_error,
-            delivered_status_ids=self._dedupe_status_ids(delivered),
-        )
-
-    async def _send_forward_chunk_to_umo(
-        self,
-        context,
-        umo: str,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        group_label: str = "",
-        header_text: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> SendOutcome:
-        nodes = self.renderer.build_nodes_for_uin(
-            10000,
-            username,
-            instance,
-            tweets,
-            start_index=tweet_start_index,
-            group_label=group_label,
-            header_text=header_text,
-            batch_summary=batch_summary,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-        attempt = await self._send_context_message(
-            context, umo, MessageChain([nodes]), "scheduled forwarded tweets"
-        )
-        if attempt.success:
-            return SendOutcome(
-                success=True,
-                delivered_status_ids=self._status_ids_from_tweets(tweets),
-            )
-        if not attempt.retryable:
-            return SendOutcome(
-                success=attempt.uncertain,
-                error=attempt.error,
-                warning=attempt.warning,
-            )
-
-        # 去掉视频后重试
-        if not media_only and any(
-            m.is_video for t in tweets for m in t.media if m.path
-        ):
-            nodes_nv = self.renderer.build_nodes_for_uin(
-                10000,
-                username,
-                instance,
-                tweets,
-                start_index=tweet_start_index,
-                exclude_videos=True,
-                group_label=group_label,
-                header_text=header_text,
-                batch_summary=batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            attempt_nv = await self._send_context_message(
-                context,
-                umo,
-                MessageChain([nodes_nv]),
-                "scheduled tweets without videos",
-            )
-            if attempt_nv.success:
-                logger.info(
-                    f"[NitterTweets] 初次失败后已向 {umo} 发送去除视频的定时推文"
-                )
-                return SendOutcome(
-                    success=not media_only,
-                    error=attempt.error,
-                    delivery_status="partial_failed",
-                    delivery_error=attempt.error,
-                    delivered_status_ids=(
-                        () if media_only else self._status_ids_from_tweets(tweets)
-                    ),
-                )
-            if not attempt_nv.retryable:
-                retry_accepted = attempt_nv.uncertain
-                return SendOutcome(
-                    success=retry_accepted and not media_only,
-                    error=attempt_nv.error or attempt.error,
-                    warning=attempt_nv.warning,
-                    delivery_status=(
-                        "partial_failed" if retry_accepted else "failed"
-                    ),
-                    delivery_error=attempt_nv.error or attempt.error,
-                    delivered_status_ids=(
-                        self._status_ids_from_tweets(tweets)
-                        if retry_accepted and not media_only
-                        else ()
-                    ),
-                )
-            attempt = attempt_nv
-
-        # Only split on explicit payload reject (retcode 1200 / res_id).
-        # Other retryable errors (timeout/network) fall through to plain text.
-        reject = bool(attempt.error) and self._is_forward_payload_rejected_error(
-            Exception(str(attempt.error))
-        )
-        remaining = tweets
-        remaining_index = tweet_start_index
-        remaining_header = header_text
-        remaining_summary = batch_summary
-        any_part_sent = False
-        delivered: list[str] = []
-        split_error = ""
-        split_warning = ""
-        if reject:
-            parts = self._split_tweets_for_forward_retry(tweets)
-            if parts:
-                logger.info(
-                    f"[NitterTweets] 定时合并转发失败，拆成 {len(parts)} 段重试 "
-                    f"(tweets={len(tweets)}, target={umo})"
-                )
-                index = tweet_start_index
-                for offset, part in enumerate(parts):
-                    part_outcome = await self._send_forward_chunk_to_umo(
-                        context,
-                        umo,
-                        username,
-                        instance,
-                        part,
-                        group_label=group_label if offset == 0 else "",
-                        header_text=header_text if offset == 0 else "",
-                        batch_summary=batch_summary if offset == 0 else "",
-                        tweet_start_index=index,
-                        media_only=media_only,
-                        omit_status_url=omit_status_url,
-                        hide_original_when_translated=hide_original_when_translated,
-                        link_style=link_style,
-                    )
-                    part_delivered = tuple(
-                        getattr(part_outcome, "delivered_status_ids", ()) or ()
-                    )
-                    if part_outcome.success and not part_delivered:
-                        part_delivered = self._status_ids_from_tweets(part)
-                    delivered.extend(part_delivered)
-                    if part_delivered:
-                        any_part_sent = True
-                    if not part_outcome.success:
-                        split_error = (
-                            getattr(part_outcome, "delivery_error", "")
-                            or part_outcome.error
-                        )
-                        split_warning = part_outcome.warning
-                        tail = []
-                        for later in parts[offset:]:
-                            tail.extend(later)
-                        delivered_set = set(self._dedupe_status_ids(delivered))
-                        remaining_with_offsets = [
-                            (tail_offset, tweet)
-                            for tail_offset, tweet in enumerate(tail)
-                            if not tweet.status_id
-                            or str(tweet.status_id) not in delivered_set
-                        ]
-                        remaining = [tweet for _, tweet in remaining_with_offsets]
-                        if remaining_with_offsets:
-                            remaining_index = index + remaining_with_offsets[0][0]
-                        keep_leading_text = offset == 0 and not any_part_sent
-                        remaining_header = header_text if keep_leading_text else ""
-                        remaining_summary = batch_summary if keep_leading_text else ""
-                        break
-                    any_part_sent = True
-                    index += len(part)
-                else:
-                    return SendOutcome(
-                        success=True,
-                        error=attempt.error,
-                        delivery_status=(
-                            "partial_failed" if attempt.error else "success"
-                        ),
-                        delivery_error=attempt.error,
-                        delivered_status_ids=self._dedupe_status_ids(delivered),
-                    )
-
-        if not remaining:
-            error = split_error or attempt.error or "forward split produced empty remainder"
-            return SendOutcome(
-                success=False,
-                error=error,
-                warning=split_warning,
-                delivery_status=(
-                    "partial_failed" if any_part_sent or delivered else "failed"
-                ),
-                delivery_error=error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-
-        # Prefer media-capable direct send for undelivered remainder only.
-        # If any split part already delivered, never re-send the full batch.
-        logger.info(
-            f"[NitterTweets] 定时合并转发仍失败，降级直发 "
-            f"(tweets={len(remaining)}/{len(tweets)}, target={umo})"
-        )
-        fallback = await self._send_direct_to_umo(
-            context,
-            umo,
-            username,
-            instance,
-            remaining,
-            group_label=group_label,
-            header_text=remaining_header,
-            batch_summary=remaining_summary,
-            tweet_start_index=remaining_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-        fallback_delivered = tuple(
-            getattr(fallback, "delivered_status_ids", ()) or ()
-        )
-        if fallback.success:
-            delivered.extend(
-                fallback_delivered or self._status_ids_from_tweets(remaining)
-            )
-            return SendOutcome(
-                success=True,
-                error=attempt.error or fallback.error or split_error,
-                warning=fallback.warning or split_warning,
-                delivery_status=(
-                    "partial_failed"
-                    if attempt.error or fallback.error or split_error or any_part_sent
-                    else "success"
-                ),
-                delivery_error=attempt.error or fallback.error or split_error or "",
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-        if fallback_delivered:
-            delivered.extend(fallback_delivered)
-            delivered_set = set(self._dedupe_status_ids(delivered))
-            remaining_with_offsets = [
-                (offset, tweet)
-                for offset, tweet in enumerate(remaining)
-                if not tweet.status_id
-                or str(tweet.status_id) not in delivered_set
-            ]
-            remaining = [tweet for _, tweet in remaining_with_offsets]
-            if remaining_with_offsets:
-                remaining_index += remaining_with_offsets[0][0]
-            remaining_header = ""
-            remaining_summary = ""
-            any_part_sent = True
-            if not remaining:
-                error = (
-                    getattr(fallback, "delivery_error", "")
-                    or fallback.error
-                    or split_error
-                )
-                return SendOutcome(
-                    success=False,
-                    error=error,
-                    warning=fallback.warning or split_warning,
-                    delivery_status="partial_failed",
-                    delivery_error=error,
-                    delivered_status_ids=self._dedupe_status_ids(delivered),
-                )
-
-        if media_only:
-            error = (
-                getattr(fallback, "delivery_error", "")
-                or fallback.error
-                or split_error
-                or attempt.error
-                or "media-only delivery failed"
-            )
-            return SendOutcome(
-                success=False,
-                error=error,
-                warning=fallback.warning or split_warning,
-                delivery_status="partial_failed" if delivered else "failed",
-                delivery_error=error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-
-        # Second-level: plain text (lighter payload) for remainder only.
-        plain = await self._send_context_message(
-            context,
-            umo,
-            MessageChain(
-                [
-                    Plain(
-                        self.renderer.format_plain(
-                            username,
-                            instance,
-                            remaining,
-                            start_index=remaining_index,
-                            group_label=group_label,
-                            header_text=remaining_header,
-                            batch_summary=remaining_summary,
-                            media_only=media_only,
-                            omit_status_url=omit_status_url,
-                            hide_original_when_translated=hide_original_when_translated,
-                            link_style=link_style,
-                        )
-                    )
-                ]
-            ),
-            "scheduled tweet plain fallback",
-        )
-        plain_ok = plain.success or plain.uncertain
-        if plain.success and not media_only:
-            delivered.extend(self._status_ids_from_tweets(remaining))
-        # Do not mark whole-batch success when only some split parts arrived:
-        # scheduler would advance seen and permanently drop the remainder.
-        # Prefer possible re-push of the first part over silent loss.
-        if plain_ok and media_only:
-            error = (
-                plain.error
-                or fallback.error
-                or split_error
-                or attempt.error
-                or "media-only fallback omitted media"
-            )
-            return SendOutcome(
-                success=False,
-                error=error,
-                warning=plain.warning or fallback.warning or split_warning,
-                delivery_status="partial_failed" if delivered else "failed",
-                delivery_error=error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-        if plain_ok:
-            return SendOutcome(
-                success=True,
-                error=attempt.error or plain.error or fallback.error or split_error,
-                warning=plain.warning or fallback.warning or split_warning,
-                delivery_status=(
-                    "partial_failed"
-                    if attempt.error
-                    or plain.error
-                    or fallback.error
-                    or split_error
-                    or any_part_sent
-                    else "success"
-                ),
-                delivery_error=(
-                    attempt.error
-                    or plain.error
-                    or fallback.error
-                    or split_error
-                    or ""
-                ),
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-        error = plain.error or fallback.error or split_error or attempt.error
-        return SendOutcome(
-            success=False,
-            error=error,
-            warning=plain.warning or fallback.warning or split_warning,
-            delivery_status=(
-                "partial_failed" if any_part_sent or delivered else "failed"
-            ),
-            delivery_error=error or "",
-            delivered_status_ids=self._dedupe_status_ids(delivered),
-        )
-
-    async def send_merged_to_umo(
-        self,
-        context,
-        umo: str,
-        batches: list[TweetBatch],
-        group_label: str = "",
-        batch_summary: str = "",
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> MergedSendOutcome:
-        adapter = self._delivery_adapter_for_umo(context, umo)
-        if link_style == "plain":
-            link_style = self.resolve_link_style(getattr(adapter, "name", ""))
-        tweet_count = self._count_batch_tweets(batches)
-        if not self._should_use_merge_for_count(tweet_count):
-            return await self._send_merged_direct_to_umo(
-                context, umo, batches, group_label, batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-
-        if not adapter.supports_merged_forward:
-            return await self._send_merged_direct_to_umo(
-                context, umo, batches, group_label, batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-
-        if self._should_chunk_forward_tweets(tweet_count):
-            return await self._send_merged_forward_chunks_to_umo(
-                context, umo, batches, group_label, batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-
-        return await self._send_merged_forward_chunk_to_umo(
-            context, umo, batches, group_label, batch_summary,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-
-    async def _send_merged_forward_chunks_to_umo(
-        self,
-        context,
-        umo: str,
-        batches: list[TweetBatch],
-        group_label: str = "",
-        batch_summary: str = "",
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> MergedSendOutcome:
-        omitted_videos = 0
-
-        def track_omitted_videos(outcome: MergedSendOutcome) -> None:
-            nonlocal omitted_videos
-            omitted_videos += outcome.omitted_videos
-
-        chunked_batches = []
-        start_index = 1
-        for chunk_index, chunk in enumerate(self._batch_chunks(batches)):
-            chunked_batches.append((chunk_index, start_index, chunk))
-            start_index += self._count_batch_tweets(chunk)
-        errors: list[str] = []
-        warnings: list[str] = []
-        delivery_errors: list[str] = []
-        delivered: list[str] = []
-        partial_delivery = False
-        for indexed_chunk in chunked_batches:
-            outcome = await self._send_merged_forward_chunk_to_umo(
-                context,
-                umo,
-                indexed_chunk[2],
-                group_label if indexed_chunk[0] == 0 else "",
-                batch_summary if indexed_chunk[0] == 0 else "",
-                tweet_start_index=indexed_chunk[1],
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            track_omitted_videos(outcome)
-            if outcome.error:
-                errors.append(outcome.error)
-            if outcome.warning:
-                warnings.append(outcome.warning)
-            status = self._normalized_delivery_status(outcome)
-            if status in {"partial_failed", "failed"}:
-                partial_delivery = True
-            outcome_delivery_error = getattr(outcome, "delivery_error", "")
-            if outcome_delivery_error:
-                delivery_errors.append(outcome_delivery_error)
-            if outcome.success:
-                delivered.extend(
-                    getattr(outcome, "delivered_status_ids", ())
-                    or self._status_ids_from_batches(indexed_chunk[2])
-                )
-            else:
-                delivered.extend(getattr(outcome, "delivered_status_ids", ()))
-                delivery_error = "; ".join(delivery_errors) or "; ".join(errors)
-                return MergedSendOutcome(
-                    success=False,
-                    mode=outcome.mode,
-                    omitted_videos=omitted_videos,
-                    error="; ".join(errors) or outcome.error,
-                    warning="; ".join(warnings),
-                    delivery_status=(
-                        "partial_failed"
-                        if delivered or status == "partial_failed"
-                        else "failed"
-                    ),
-                    delivery_error=delivery_error or outcome.error,
-                    delivered_status_ids=self._dedupe_status_ids(delivered),
-                )
-        final_delivery_error = "; ".join(delivery_errors)
-        if partial_delivery and not final_delivery_error:
-            final_delivery_error = "; ".join(errors)
-        return MergedSendOutcome(
-            success=True,
-            mode="chunked_forward",
-            omitted_videos=omitted_videos,
-            error="; ".join(errors),
-            warning="; ".join(warnings),
-            delivery_status="partial_failed" if partial_delivery else "success",
-            delivery_error=final_delivery_error,
-            delivered_status_ids=self._dedupe_status_ids(delivered),
-        )
-
-    async def _send_merged_forward_chunk_to_umo(
-        self,
-        context,
-        umo: str,
-        batches: list[TweetBatch],
-        group_label: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> MergedSendOutcome:
-        omitted_videos = self._count_attached_videos(batches)
-        has_video = self._merged_forward_has_video(batches)
-        raw_forward_available = (
-            has_video and self._onebot_call_action_for_umo(context, umo) is not None
-        )
-        attempt = SendAttempt(
-            success=False,
-            retryable=True,
-            error="video merged forward not attempted",
-        )
-        if raw_forward_available:
-            raw_nodes = self.renderer.build_merged_onebot_nodes_for_uin(
-                10000,
-                batches,
-                start_index=tweet_start_index,
-                group_label=group_label,
-                batch_summary=batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            attempt = await self._send_onebot_umo_forward(
-                context, umo, raw_nodes, "merged scheduled tweets"
-            )
-            if attempt.success:
-                return MergedSendOutcome(success=True, mode="raw_forward")
-            if not attempt.retryable:
-                return MergedSendOutcome(
-                    success=attempt.uncertain,
-                    mode="uncertain_delivery" if attempt.uncertain else "failed",
-                    omitted_videos=omitted_videos,
-                    error=attempt.error,
-                    warning=attempt.warning,
-                )
-
-        if not raw_forward_available:
-            nodes = self.renderer.build_merged_nodes_for_uin(
-                10000,
-                batches,
-                start_index=tweet_start_index,
-                group_label=group_label,
-                batch_summary=batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            attempt = await self._send_context_message(
-                context, umo, MessageChain([nodes]), "merged scheduled tweets"
-            )
-            if attempt.success:
-                return MergedSendOutcome(success=True, mode="full_forward")
-            if not attempt.retryable:
-                return MergedSendOutcome(
-                    success=attempt.uncertain,
-                    mode="uncertain_delivery" if attempt.uncertain else "failed",
-                    omitted_videos=omitted_videos,
-                    error=attempt.error,
-                    warning=attempt.warning,
-                )
-
-        if omitted_videos and not media_only:
-            raw_nodes_nv = self.renderer.build_merged_onebot_nodes_for_uin(
-                10000,
-                batches,
-                start_index=tweet_start_index,
-                exclude_videos=True,
-                group_label=group_label,
-                batch_summary=batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            raw_retry_attempt = await self._send_onebot_umo_forward(
-                context,
-                umo,
-                raw_nodes_nv,
-                "merged tweets without videos",
-            )
-            if raw_retry_attempt.success:
-                logger.warning(
-                    f"[NitterTweets] 初次失败后已向 {umo} 发送去除 "
-                    f"{omitted_videos} 个视频/GIF 附件的合并推文"
-                )
-                return MergedSendOutcome(
-                    success=not media_only,
-                    mode="raw_forward_without_videos",
-                    omitted_videos=omitted_videos,
-                    error=attempt.error,
-                    delivery_status="partial_failed",
-                    delivery_error=attempt.error,
-                )
-            if not raw_retry_attempt.retryable:
-                retry_accepted = raw_retry_attempt.uncertain
-                return MergedSendOutcome(
-                    success=retry_accepted and not media_only,
-                    mode=(
-                        "uncertain_delivery" if retry_accepted else "failed"
-                    ),
-                    omitted_videos=omitted_videos,
-                    error=raw_retry_attempt.error or attempt.error,
-                    warning=raw_retry_attempt.warning,
-                    delivery_status=(
-                        "partial_failed" if retry_accepted else "failed"
-                    ),
-                    delivery_error=raw_retry_attempt.error or attempt.error,
-                )
-            nodes_nv = self.renderer.build_merged_nodes_for_uin(
-                10000,
-                batches,
-                start_index=tweet_start_index,
-                exclude_videos=True,
-                group_label=group_label,
-                batch_summary=batch_summary,
-                media_only=media_only,
-                omit_status_url=omit_status_url,
-                hide_original_when_translated=hide_original_when_translated,
-                link_style=link_style,
-            )
-            retry_attempt = await self._send_context_message(
-                context,
-                umo,
-                MessageChain([nodes_nv]),
-                "merged tweets without videos",
-            )
-            if retry_attempt.success:
-                logger.warning(
-                    f"[NitterTweets] 初次失败后已向 {umo} 发送去除 "
-                    f"{omitted_videos} 个视频/GIF 附件的合并推文"
-                )
-                return MergedSendOutcome(
-                    success=not media_only,
-                    mode="forward_without_videos",
-                    omitted_videos=omitted_videos,
-                    error=attempt.error,
-                    delivery_status="partial_failed",
-                    delivery_error=attempt.error,
-                )
-            if not retry_attempt.retryable:
-                retry_accepted = retry_attempt.uncertain
-                return MergedSendOutcome(
-                    success=retry_accepted and not media_only,
-                    mode=(
-                        "uncertain_delivery" if retry_accepted else "failed"
-                    ),
-                    omitted_videos=omitted_videos,
-                    error=retry_attempt.error or attempt.error,
-                    warning=retry_attempt.warning,
-                    delivery_status=(
-                        "partial_failed" if retry_accepted else "failed"
-                    ),
-                    delivery_error=retry_attempt.error or attempt.error,
-                )
-            attempt = retry_attempt
-
-        reject = bool(attempt.error) and self._is_forward_payload_rejected_error(
-            Exception(str(attempt.error))
-        )
-        remaining_batches = batches
-        remaining_start_index = tweet_start_index
-        remaining_group_label = group_label
-        remaining_summary = batch_summary
-        delivered: list[str] = []
-        split_error = ""
-        split_warning = ""
-        split_omitted_videos = 0
-        if reject:
-            parts = self._split_batches_for_forward_retry(batches)
-            if parts:
-                logger.info(
-                    f"[NitterTweets] 定时多批次合并转发失败，拆成 {len(parts)} 段重试 "
-                    f"(tweets={self._count_batch_tweets(batches)}, target={umo})"
-                )
-                index = tweet_start_index
-                for offset, part in enumerate(parts):
-                    part_outcome = await self._send_merged_forward_chunk_to_umo(
-                        context,
-                        umo,
-                        part,
-                        group_label if offset == 0 else "",
-                        batch_summary if offset == 0 else "",
-                        tweet_start_index=index,
-                        media_only=media_only,
-                        omit_status_url=omit_status_url,
-                        hide_original_when_translated=hide_original_when_translated,
-                        link_style=link_style,
-                    )
-                    split_omitted_videos += part_outcome.omitted_videos
-                    part_delivered = tuple(
-                        getattr(part_outcome, "delivered_status_ids", ()) or ()
-                    )
-                    if part_outcome.success and not part_delivered:
-                        part_delivered = self._status_ids_from_batches(part)
-                    delivered.extend(part_delivered)
-                    if not part_outcome.success:
-                        split_error = (
-                            getattr(part_outcome, "delivery_error", "")
-                            or part_outcome.error
-                        )
-                        split_warning = part_outcome.warning
-                        tail = []
-                        for later in parts[offset:]:
-                            tail.extend(later)
-                        (
-                            remaining_batches,
-                            first_remaining_offset,
-                        ) = self._batches_without_status_ids(tail, delivered)
-                        remaining_start_index = index + first_remaining_offset
-                        keep_leading_text = offset == 0 and not delivered
-                        remaining_group_label = (
-                            group_label if keep_leading_text else ""
-                        )
-                        remaining_summary = (
-                            batch_summary if keep_leading_text else ""
-                        )
-                        break
-                    index += self._count_batch_tweets(part)
-                else:
-                    return MergedSendOutcome(
-                        success=True,
-                        mode="split_forward",
-                        omitted_videos=split_omitted_videos,
-                        error=attempt.error,
-                        delivery_status="partial_failed",
-                        delivery_error=attempt.error,
-                        delivered_status_ids=self._dedupe_status_ids(delivered),
-                    )
-
-        if not remaining_batches:
-            error = split_error or attempt.error or "forward split produced empty remainder"
-            return MergedSendOutcome(
-                success=False,
-                mode="failed",
-                omitted_videos=split_omitted_videos or omitted_videos,
-                error=error,
-                warning=split_warning,
-                delivery_status="partial_failed" if delivered else "failed",
-                delivery_error=error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-
-        if media_only:
-            error = split_error or attempt.error or "media-only delivery failed"
-            return MergedSendOutcome(
-                success=False,
-                mode="failed",
-                omitted_videos=split_omitted_videos or omitted_videos,
-                error=error,
-                warning=split_warning,
-                delivery_status="partial_failed" if delivered else "failed",
-                delivery_error=error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-
-        fallback = await self._send_context_message(
-            context,
-            umo,
-            MessageChain(
-                [
-                    Plain(
-                        self.renderer.format_merged_plain(
-                            remaining_batches,
-                            start_index=remaining_start_index,
-                            group_label=remaining_group_label,
-                            batch_summary=remaining_summary,
-                            media_only=media_only,
-                            omit_status_url=omit_status_url,
-                            hide_original_when_translated=hide_original_when_translated,
-                            link_style=link_style,
-                        )
-                    )
-                ]
-            ),
-            "merged scheduled tweet fallback",
-        )
-        if fallback.success or fallback.uncertain:
-            if media_only:
-                error = (
-                    split_error
-                    or attempt.error
-                    or fallback.error
-                    or "media-only fallback omitted media"
-                )
-                return MergedSendOutcome(
-                    success=False,
-                    mode="plain_fallback",
-                    omitted_videos=split_omitted_videos or omitted_videos,
-                    error=error,
-                    warning=fallback.warning or split_warning,
-                    delivery_status="partial_failed" if delivered else "failed",
-                    delivery_error=error,
-                    delivered_status_ids=self._dedupe_status_ids(delivered),
-                )
-            if fallback.success:
-                delivered.extend(self._status_ids_from_batches(remaining_batches))
-            return MergedSendOutcome(
-                success=True,
-                mode="uncertain_delivery" if fallback.uncertain else "plain_fallback",
-                omitted_videos=split_omitted_videos or omitted_videos,
-                error=split_error or attempt.error,
-                warning=fallback.warning or split_warning,
-                delivery_status=(
-                    "partial_failed"
-                    if split_error or attempt.error or delivered
-                    else "success"
-                ),
-                delivery_error=split_error or attempt.error,
-                delivered_status_ids=self._dedupe_status_ids(delivered),
-            )
-        error = fallback.error or split_error or attempt.error
-        return MergedSendOutcome(
-            success=False,
-            mode="failed",
-            omitted_videos=split_omitted_videos or omitted_videos,
-            error=error,
-            warning=fallback.warning or split_warning,
-            delivery_status="partial_failed" if delivered else "failed",
-            delivery_error=error,
-            delivered_status_ids=self._dedupe_status_ids(delivered),
-        )
-
-    async def _send_direct_event(
-        self,
-        event,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        notices: list[str] | None = None,
-        header_text: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> bool:
-        return await self._delivery_adapter_for_event(event).send_event(
-            event,
-            username,
-            instance,
-            tweets,
-            notices=notices,
-            header_text=header_text,
-            tweet_start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-
-    async def _send_default_direct_event(
-        self,
-        event,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        notices: list[str] | None = None,
-        header_text: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> bool:
-        profile = self.platform_resolver.from_event(event)
-        return await DefaultDeliveryAdapter(self, profile).send_event(
-            event,
-            username,
-            instance,
-            tweets,
-            notices=notices,
-            header_text=header_text,
-            tweet_start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-
-    async def _send_direct_to_umo(
-        self,
-        context,
-        umo: str,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        group_label: str = "",
-        header_text: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> SendOutcome:
-        return await self._delivery_adapter_for_umo(context, umo).send_to_umo(
-            context,
-            umo,
-            username,
-            instance,
-            tweets,
-            group_label=group_label,
-            header_text=header_text,
-            batch_summary=batch_summary,
-            tweet_start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-
-    async def _send_default_direct_to_umo(
-        self,
-        context,
-        umo: str,
-        username: str,
-        instance: str,
-        tweets: list[TweetItem],
-        group_label: str = "",
-        header_text: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> SendOutcome:
-        profile = self.platform_resolver.from_umo(context, umo)
-        return await DefaultDeliveryAdapter(self, profile).send_to_umo(
-            context,
-            umo,
-            username,
-            instance,
-            tweets,
-            group_label=group_label,
-            header_text=header_text,
-            batch_summary=batch_summary,
-            tweet_start_index=tweet_start_index,
-            media_only=media_only,
-            omit_status_url=omit_status_url,
-            hide_original_when_translated=hide_original_when_translated,
-            link_style=link_style,
-        )
-
-    async def _send_merged_direct_to_umo(
-        self,
-        context,
-        umo: str,
-        batches: list[TweetBatch],
-        group_label: str = "",
-        batch_summary: str = "",
-        tweet_start_index: int = 1,
-        media_only: bool = False,
-        omit_status_url: bool = True,
-        hide_original_when_translated: bool = False,
-        link_style: str = "plain",
-    ) -> MergedSendOutcome:
-        omitted_videos = self._count_attached_videos(batches)
-        attempt = await self._send_context_message(
-            context,
-            umo,
-            MessageChain(
-                self.renderer.build_merged_direct_components(
-                    batches,
-                    start_index=tweet_start_index,
-                    group_label=group_label,
-                    batch_summary=batch_summary,
-                    media_only=media_only,
-                    omit_status_url=omit_status_url,
-                    hide_original_when_translated=hide_original_when_translated,
-                    link_style=link_style,
-                )
-            ),
-            "direct merged tweets",
-        )
-        if attempt.success:
-            return MergedSendOutcome(success=True, mode="direct_message")
-        if not attempt.retryable:
-            return MergedSendOutcome(
-                success=attempt.uncertain,
-                mode="uncertain_delivery" if attempt.uncertain else "failed",
-                omitted_videos=omitted_videos,
-                error=attempt.error,
-                warning=attempt.warning,
-            )
-
-        if media_only:
-            return MergedSendOutcome(
-                success=False,
-                mode="failed",
-                omitted_videos=omitted_videos,
-                error=attempt.error,
-                warning=attempt.warning,
-                delivery_status="failed",
-                delivery_error=attempt.error,
-            )
-
-        if omitted_videos:
-            retry_attempt = await self._send_context_message(
-                context,
-                umo,
-                MessageChain(
-                    self.renderer.build_merged_direct_components(
-                        batches,
-                        start_index=tweet_start_index,
-                        exclude_videos=True,
-                        group_label=group_label,
-                        batch_summary=batch_summary,
-                        media_only=media_only,
-                        omit_status_url=omit_status_url,
-                        hide_original_when_translated=hide_original_when_translated,
-                        link_style=link_style,
-                    )
-                ),
-                "direct merged tweets without videos",
-            )
-            if retry_attempt.success:
-                logger.warning(
-                    f"[NitterTweets] 初次失败后已向 {umo} 发送去除 "
-                    f"{omitted_videos} 个视频/GIF 附件的直发合并推文"
-                )
-                return MergedSendOutcome(
-                    success=not media_only,
-                    mode="direct_without_videos",
-                    omitted_videos=omitted_videos,
-                    error=attempt.error,
-                    delivery_status="partial_failed",
-                    delivery_error=attempt.error,
-                )
-            if not retry_attempt.retryable:
-                retry_accepted = retry_attempt.uncertain
-                return MergedSendOutcome(
-                    success=retry_accepted and not media_only,
-                    mode=(
-                        "uncertain_delivery" if retry_accepted else "failed"
-                    ),
-                    omitted_videos=omitted_videos,
-                    error=retry_attempt.error or attempt.error,
-                    warning=retry_attempt.warning,
-                    delivery_status=(
-                        "partial_failed" if retry_accepted else "failed"
-                    ),
-                    delivery_error=retry_attempt.error or attempt.error,
-                )
-            attempt = retry_attempt
-
-        fallback = await self._send_context_message(
-            context,
-            umo,
-            MessageChain(
-                [
-                    Plain(
-                        self.renderer.format_merged_plain(
-                            batches,
-                            start_index=tweet_start_index,
-                            group_label=group_label,
-                            batch_summary=batch_summary,
-                            media_only=media_only,
-                            omit_status_url=omit_status_url,
-                            hide_original_when_translated=hide_original_when_translated,
-                            link_style=link_style,
-                        )
-                    )
-                ]
-            ),
-            "direct merged fallback",
-        )
-        if fallback.success or fallback.uncertain:
-            if media_only:
-                error = attempt.error or fallback.error or "media-only fallback omitted media"
-                return MergedSendOutcome(
-                    success=False,
-                    mode="plain_fallback",
-                    omitted_videos=omitted_videos,
-                    error=error,
-                    warning=fallback.warning,
-                    delivery_status="failed",
-                    delivery_error=error,
-                )
-            return MergedSendOutcome(
-                success=True,
-                mode="uncertain_delivery" if fallback.uncertain else "plain_fallback",
-                omitted_videos=omitted_videos,
-                error=attempt.error,
-                warning=fallback.warning,
-                delivery_status="partial_failed" if attempt.error else "success",
-                delivery_error=attempt.error,
-            )
-        return MergedSendOutcome(
-            success=False,
-            mode="failed",
-            omitted_videos=omitted_videos,
-            error=fallback.error or attempt.error,
         )
 
     async def _send_context_message(
@@ -1858,95 +425,6 @@ class TweetSender:
         return False
 
     @classmethod
-    def event_target(cls, event) -> str:
-        return cls._event_target(event)
-
-    @classmethod
-    def _event_target(cls, event) -> str:
-        try:
-            umo = getattr(event, "unified_msg_origin", "")
-        except Exception:
-            umo = ""
-        if umo:
-            return str(umo)
-
-        group_id = safe_call(event, "get_group_id")
-        if group_id:
-            return f"group:{group_id}"
-
-        sender_id = safe_call(event, "get_sender_id")
-        if sender_id:
-            return f"private:{sender_id}"
-
-        platform = cls._event_platform(event)
-        return platform or "unknown"
-
-    @staticmethod
-    def _count_attached_videos(batches: list[TweetBatch]) -> int:
-        return sum(
-            1
-            for _, _, tweets in batches
-            for tweet in tweets
-            for media in tweet.media
-            if media.path and media.is_video
-        )
-
-    @staticmethod
-    def _count_batch_tweets(batches: list[TweetBatch]) -> int:
-        return sum(len(tweets) for _, _, tweets in batches)
-
-    @staticmethod
-    def _has_attached_videos(tweets: list[TweetItem]) -> bool:
-        return any(
-            media.is_video for tweet in tweets for media in tweet.media if media.path
-        )
-
-    @staticmethod
-    def _has_attached_images(tweets: list[TweetItem]) -> bool:
-        return any(
-            media.is_image for tweet in tweets for media in tweet.media if media.path
-        )
-
-    def _should_split_qq_direct_videos(self, event, tweets: list[TweetItem]) -> bool:
-        return bool(
-            self.send_video_attachments
-            and self._has_attached_videos(tweets)
-            and self._should_split_direct_videos_for_event(event)
-        )
-
-    def _should_split_qq_direct_videos_for_umo(
-        self,
-        context,
-        umo: str,
-        tweets: list[TweetItem],
-    ) -> bool:
-        return bool(
-            self.send_video_attachments
-            and self._has_attached_videos(tweets)
-            and self._should_split_direct_videos_for_umo(context, umo)
-        )
-
-    def _merged_forward_has_video(self, batches: list[TweetBatch]) -> bool:
-        return bool(
-            self.send_video_attachments
-            and any(
-                media.path and media.is_video
-                for _, _, tweets in batches
-                for tweet in tweets
-                for media in tweet.media
-            )
-        )
-
-    def _should_use_merge_for_count(self, tweet_count: int) -> bool:
-        return (
-            self.merge_tweet_threshold > 0
-            and tweet_count >= self.merge_tweet_threshold
-        )
-
-    def _should_chunk_forward_tweets(self, tweet_count: int) -> bool:
-        return tweet_count > self.FORWARD_TWEET_CHUNK_SIZE
-
-    @classmethod
     def _is_forward_payload_rejected_error(cls, exc: Exception | None) -> bool:
         """OneBot/NapCat explicit reject of merged-forward content (e.g. retcode 1200)."""
         if exc is None:
@@ -1967,287 +445,3 @@ class TweetSender:
         if "发送转发消息" in text and "失败" in text:
             return True
         return False
-
-    def _split_tweets_for_forward_retry(
-        self, tweets: list[TweetItem]
-    ) -> list[list[TweetItem]] | None:
-        if len(tweets) <= self.FORWARD_SPLIT_MIN_TWEETS:
-            return None
-        mid = max(1, len(tweets) // 2)
-        left = tweets[:mid]
-        right = tweets[mid:]
-        if not left or not right:
-            return None
-        return [left, right]
-
-    @staticmethod
-    def _notify_sent_progress(callback, count: int) -> None:
-        if not callable(callback):
-            return
-        try:
-            callback(count)
-        except Exception as exc:
-            logger.warning(f"[NitterTweets] 手动发送进度回调失败: {exc}")
-
-    @staticmethod
-    def _notify_delivered(callback, count: int) -> None:
-        if not callable(callback) or count <= 0:
-            return
-        try:
-            callback(count)
-        except Exception as exc:
-            logger.warning(f"[NitterTweets] 手动发送分段进度回调失败: {exc}")
-
-    @staticmethod
-    async def _send_chunked_bool(chunks, send_chunk) -> bool:
-        for chunk in chunks:
-            if not await send_chunk(chunk):
-                return False
-        return True
-
-    @staticmethod
-    async def _send_chunked_outcomes(
-        chunks,
-        send_chunk,
-        success_factory,
-        failure_factory,
-        after_outcome=None,
-    ):
-        errors: list[str] = []
-        warnings: list[str] = []
-        for chunk in chunks:
-            outcome = await send_chunk(chunk)
-            if after_outcome is not None:
-                after_outcome(outcome)
-            if outcome.error:
-                errors.append(outcome.error)
-            if outcome.warning:
-                warnings.append(outcome.warning)
-            if not outcome.success:
-                return failure_factory(
-                    outcome,
-                    "; ".join(errors),
-                    "; ".join(warnings),
-                )
-        return success_factory("; ".join(errors), "; ".join(warnings))
-
-    def _tweet_chunks(self, tweets: list[TweetItem]) -> list[list[TweetItem]]:
-        size = self.FORWARD_TWEET_CHUNK_SIZE
-        return [tweets[index : index + size] for index in range(0, len(tweets), size)]
-
-    @staticmethod
-    def _status_ids_from_tweets(tweets: list[TweetItem]) -> tuple[str, ...]:
-        return tuple(
-            str(getattr(tweet, "status_id", "") or "")
-            for tweet in tweets
-            if str(getattr(tweet, "status_id", "") or "")
-        )
-
-    @classmethod
-    def _status_ids_from_batches(
-        cls, batches: list[TweetBatch]
-    ) -> tuple[str, ...]:
-        ids: list[str] = []
-        for _username, _instance, tweets in batches:
-            ids.extend(cls._status_ids_from_tweets(tweets))
-        return tuple(ids)
-
-    @staticmethod
-    def _dedupe_status_ids(values) -> tuple[str, ...]:
-        seen: set[str] = set()
-        result: list[str] = []
-        for value in values or ():
-            text = str(value or "")
-            if text and text not in seen:
-                seen.add(text)
-                result.append(text)
-        return tuple(result)
-
-    @staticmethod
-    def _normalized_delivery_status(outcome) -> str:
-        return str(
-            getattr(outcome, "delivery_status", "success") or "success"
-        ).strip().lower()
-
-    def _batch_chunks(self, batches: list[TweetBatch]) -> list[list[TweetBatch]]:
-        chunks: list[list[TweetBatch]] = []
-        current: list[TweetBatch] = []
-        current_count = 0
-        size = self.FORWARD_TWEET_CHUNK_SIZE
-
-        for username, instance, tweets in batches:
-            for tweet_chunk in self._tweet_chunks(tweets):
-                if current and current_count + len(tweet_chunk) > size:
-                    chunks.append(current)
-                    current = []
-                    current_count = 0
-                current.append((username, instance, tweet_chunk))
-                current_count += len(tweet_chunk)
-
-        if current:
-            chunks.append(current)
-        return chunks
-
-    def _split_batches_for_forward_retry(
-        self, batches: list[TweetBatch]
-    ) -> list[list[TweetBatch]] | None:
-        total = self._count_batch_tweets(batches)
-        if total <= self.FORWARD_SPLIT_MIN_TWEETS:
-            return None
-        mid = max(1, total // 2)
-        left: list[TweetBatch] = []
-        right: list[TweetBatch] = []
-        consumed = 0
-        for username, instance, tweets in batches:
-            split_at = max(0, min(len(tweets), mid - consumed))
-            if split_at:
-                left.append((username, instance, list(tweets[:split_at])))
-            if split_at < len(tweets):
-                right.append((username, instance, list(tweets[split_at:])))
-            consumed += split_at
-        if not left or not right:
-            return None
-        return [left, right]
-
-    @staticmethod
-    def _batches_without_status_ids(
-        batches: list[TweetBatch], delivered_status_ids
-    ) -> tuple[list[TweetBatch], int]:
-        delivered = {str(value or "") for value in delivered_status_ids if value}
-        result: list[TweetBatch] = []
-        first_remaining_offset: int | None = None
-        offset = 0
-        for username, instance, tweets in batches:
-            remaining = []
-            for tweet in tweets:
-                status_id = str(getattr(tweet, "status_id", "") or "")
-                if status_id and status_id in delivered:
-                    offset += 1
-                    continue
-                if first_remaining_offset is None:
-                    first_remaining_offset = offset
-                remaining.append(tweet)
-                offset += 1
-            if remaining:
-                result.append((username, instance, remaining))
-        return result, first_remaining_offset or 0
-
-    def supports_merged_forward_for_umo(self, context, umo: str) -> bool:
-        return self._delivery_adapter_for_umo(context, umo).supports_merged_forward
-
-    def _delivery_adapter_for_umo(self, context, umo: str):
-        profile = self.platform_resolver.from_umo(context, umo)
-        return self.delivery_registry.adapter_for(self, profile)
-
-    def _delivery_adapter_for_event(self, event):
-        profile = self.platform_resolver.from_event(event)
-        return self.delivery_registry.adapter_for(self, profile)
-
-    @classmethod
-    def _should_split_direct_videos_for_umo(cls, context, umo: str) -> bool:
-        profile = PlatformResolver().from_umo(context, umo)
-        return profile.should_split_qq_direct_videos
-
-    @classmethod
-    def _should_split_direct_videos_for_event(cls, event) -> bool:
-        profile = PlatformResolver().from_event(event)
-        return profile.should_split_qq_direct_videos
-
-    @classmethod
-    def _should_use_lark_for_umo(cls, context, umo: str) -> bool:
-        return PlatformResolver().from_umo(context, umo).is_lark
-
-    @classmethod
-    def _should_use_lark_for_event(cls, event) -> bool:
-        return PlatformResolver().from_event(event).is_lark
-
-    @classmethod
-    def _is_lark_platform(cls, platform: str) -> bool:
-        return is_lark_platform(platform)
-
-    @classmethod
-    def _should_use_forward_for_umo(cls, context, umo: str) -> bool:
-        return PlatformResolver().from_umo(context, umo).is_onebot
-
-    @classmethod
-    def _should_use_forward_for_event(cls, event) -> bool:
-        return PlatformResolver().from_event(event).is_onebot
-
-    @classmethod
-    def _is_forward_platform(cls, platform: str) -> bool:
-        return normalize_platform(platform) in cls.FORWARD_MESSAGE_PLATFORMS
-
-    @classmethod
-    def _is_qq_direct_video_split_platform(cls, platform: str) -> bool:
-        return normalize_platform(platform) in cls.QQ_DIRECT_VIDEO_SPLIT_PLATFORMS
-
-    @staticmethod
-    def _normalize_platform(platform: str) -> str:
-        return normalize_platform(platform)
-
-    @staticmethod
-    def _platform_from_umo(umo: str) -> str:
-        return str(umo or "").split(":", 1)[0].strip()
-
-    @classmethod
-    def _platform_inst_from_context(cls, context, platform_id: str):
-        return PlatformResolver().platform_inst_from_context(context, platform_id)
-
-    @classmethod
-    def _platform_type_from_context(cls, context, platform_id: str) -> str:
-        profile = PlatformResolver().from_umo(context, f"{platform_id}:GroupMessage:")
-        return profile.platform_types[0] if profile.platform_types else ""
-
-    @staticmethod
-    def _safe_platform_meta(platform):
-        return PlatformResolver.safe_platform_meta(platform)
-
-    @classmethod
-    def _event_platform(cls, event) -> str:
-        profile = PlatformResolver().from_event(event)
-        return profile.platform_id or (profile.platform_types[0] if profile.platform_types else "")
-
-    async def _send_onebot_forward(self, event, raw_nodes: list[dict]) -> bool:
-        send_forward = getattr(
-            self._delivery_adapter_for_event(event), "send_event_forward", None
-        )
-        if not callable(send_forward):
-            return False
-        return await send_forward(event, raw_nodes)
-
-    async def _send_onebot_umo_forward(
-        self,
-        context,
-        umo: str,
-        raw_nodes: list[dict],
-        label: str,
-    ) -> SendAttempt:
-        send_forward = getattr(
-            self._delivery_adapter_for_umo(context, umo), "send_umo_forward", None
-        )
-        if not callable(send_forward):
-            return SendAttempt(
-                success=False,
-                retryable=True,
-                error="OneBot call_action unavailable for proactive merged forward",
-            )
-        return await send_forward(context, umo, raw_nodes, label)
-
-    @classmethod
-    def _onebot_call_action_for_umo(cls, context, umo: str):
-        return PlatformResolver().from_umo(context, umo).call_action
-
-    @staticmethod
-    def _onebot_target_from_umo(umo: str) -> tuple[str, int | str]:
-        return OneBotDeliveryAdapter.onebot_target_from_umo(umo)
-
-    @staticmethod
-    async def _call_forward_action(
-        call_action,
-        action: str,
-        base_payload: dict,
-        raw_nodes: list[dict],
-    ) -> None:
-        await OneBotDeliveryAdapter.call_forward_action(
-            call_action, action, base_payload, raw_nodes
-        )
